requirements.txt
flask==3.0.0
flask-cors==4.0.0
supabase==2.3.5
pydantic==2.6.1
PyJWT==2.8.0
python-dotenv==1.0.0
urllib3==2.2.2
requests==2.31.0
httpx>=0.24,<0.26
flask-limiter==3.5.0
structlog==24.1.0
python-jose==3.3.0
gunicorn==21.2.0
flask-restful==0.3.10
flask-jwt-extended==4.6.0 